requirements.txt
mcp>=1.14.1
requests>=2.32.5
python-dotenv>=1.1.1
asyncio>=4.0.0
websockets>=15.0.1
pydantic>=2.11.0
httpx>=0.27.1
uvicorn>=0.31.1
aiofiles>=23.2.1
pillow>=10.0.0
python-dateutil>=2.8.2
